+++ SuperHopper/docs/release-notes.md
@@ -1,6 +1,10 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* The super hopper now transfers every possible item every 10 minutes, instead of one stack per 10 minutes.
+* Fixed super hopper collecting input if the chest below is full.
+
 ## 1.0.1
 Released 11 September 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild.